Trigger event handlers when syncing channel
Found by: wilkowy
Patch by: michaelortmann
Fixes: #599

When the bot joins a chan, gotjoin() creates a member for the channel member list and then it resets the chan with flags CHAN_RESETALL, removing that member again. This PR unsets flag CHAN_RESETWHO for this very case, when the member list contains 1 member, which is the bot itself that just got created.